lens condition deleted and renamed context
