+++ src/App.tsx
@@ -1,12 +1,12 @@
 import "./App.css";
 import { Conditions } from "./components/content/conditions";
 import { CTWProvider } from "./components/core/ctw-provider";
-import { IDProvider } from "./components/core/id-provider";
+import { IDProvider } from "./components/core/patient-provider";
 import { SYSTEM_HEALTHIE_ID } from "./fhir/system-urls";
 import "./styles/tailwind-gen.css";
 
 const AUTH_TOKEN = import.meta.env.VITE_AUTH_TOKEN;
-const SB_SARAH_ID = "000002";
+const SB_SARAH_ID = "000001";
 
 function App() {
   return (
